- Expicitly cast parameters before passing into NSString formatters assuming 64-bit lengths; this addresses exceptions and hangs viewing table contents on 32-bit machines, many thanks to stuart02 for identifying and tracking down this issue

+++ Frameworks/MCPKit/MCPFoundationKit/MCPResult.m
@@ -746,7 +746,7 @@ const OUR_CHARSET our_charsets60[] =
 		NSMutableDictionary *fieldStructure = [NSMutableDictionary dictionaryWithCapacity:39];
 
 		/* Original column position */
-		[fieldStructure setObject:[NSString stringWithFormat:@"%llu", i] forKey:@"datacolumnindex"];
+		[fieldStructure setObject:[NSString stringWithFormat:@"%llu", (unsigned long long)i] forKey:@"datacolumnindex"];
 
 		/* Name of column */
 		[fieldStructure setObject:[self stringWithCString:theField[i].name] forKey:@"name"];
